requirements.txt
asgiref==3.7.2
azure-core==1.30.1
azure-storage-blob==12.20.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
Django==4.2.6
idna==3.4
isodate==0.6.1
pycparser==2.22
python-decouple==3.8
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.1.0
whitenoise==6.6.0
